% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colt-package.R
\docType{package}
\name{colt-package}
\alias{colt}
\alias{colt-package}
\title{Colt - CLI color themes for R}
\description{
Colt is aimed at package developers that want to create themeable command
line output. It is a simple wrapper around the crayon package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s-fleck/colt}
  \item Report bugs at \url{https://github.com/s-fleck/colt/issues}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com}

}
\keyword{internal}
