% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{colt_styles}
\alias{colt_styles}
\alias{clt_default}
\alias{clt_h1}
\alias{clt_h2}
\alias{clt_chr}
\alias{clt_chr_accent}
\alias{clt_chr_subtle}
\alias{clt_true}
\alias{clt_false}
\alias{clt_maybe}
\alias{clt_num}
\alias{clt_num_accent}
\alias{clt_num_subtle}
\alias{clt_bg}
\alias{clt_bg_subtle}
\alias{clt_bg_accent}
\alias{clt_na}
\alias{clt_nan}
\alias{clt_inf}
\alias{clt_null}
\alias{clt_info}
\alias{clt_message}
\alias{clt_warning}
\alias{clt_error}
\alias{clt_emph}
\alias{clt_emph2}
\title{Colt style elements}
\usage{
clt_default(x, fun = options("colt.theme")[[1]]$clt_default)

clt_h1(x, fun = options("colt.theme")[[1]]$clt_h1)

clt_h2(x, fun = options("colt.theme")[[1]]$clt_h2)

clt_chr(x, fun = options("colt.theme")[[1]]$clt_chr)

clt_chr_accent(x, fun = options("colt.theme")[[1]]$clt_chr_accent)

clt_chr_subtle(x, fun = options("colt.theme")[[1]]$clt_chr_subtle)

clt_true(x, fun = options("colt.theme")[[1]]$clt_true)

clt_false(x, fun = options("colt.theme")[[1]]$clt_false)

clt_maybe(x, fun = options("colt.theme")[[1]]$clt_maybe)

clt_num(x, fun = options("colt.theme")[[1]]$clt_num)

clt_num_accent(x, fun = options("colt.theme")[[1]]$clt_num_accent)

clt_num_subtle(x, fun = options("colt.theme")[[1]]$clt_num_subtle)

clt_bg(x, fun = options("colt.theme")[[1]]$clt_bg)

clt_bg_subtle(x, fun = options("colt.theme")[[1]]$clt_bg_subtle)

clt_bg_accent(x, fun = options("colt.theme")[[1]]$clt_bg_accent)

clt_na(x, fun = options("colt.theme")[[1]]$clt_na)

clt_nan(x, fun = options("colt.theme")[[1]]$clt_nan)

clt_inf(x, fun = options("colt.theme")[[1]]$clt_inf)

clt_null(x, fun = options("colt.theme")[[1]]$clt_null)

clt_info(x, fun = options("colt.theme")[[1]]$clt_info)

clt_message(x, fun = options("colt.theme")[[1]]$clt_message)

clt_warning(x, fun = options("colt.theme")[[1]]$clt_warning)

clt_error(x, fun = options("colt.theme")[[1]]$clt_error)

clt_emph(x, fun = options("colt.theme")[[1]]$clt_emph)

clt_emph2(x, fun = options("colt.theme")[[1]]$clt_emph2)
}
\arguments{
\item{x}{a character vector}

\item{fun}{a styling function}
}
\description{
Colt style elements
}
