% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colt_theme.R
\name{print.colt_theme}
\alias{print.colt_theme}
\title{Printing Colt Themes}
\usage{
\method{print}{colt_theme}(x, ...)
}
\arguments{
\item{x}{a \code{colt_theme}}

\item{...}{ignored}
}
\description{
Printing Colt Themes
}
